fix: mypy 型エラーを修正
修正内容:
- tests/data を mypy 除外対象に追加
- flask_util.py テスト: cache_control=None を "" に変更
- pil_util.py: draw_text 戻り値型を tuple[float, float] に修正
- panel_util.py: next_pos_y を float 型に明示
- selenium_util.py: _reap_chrome_processes の呼び出し名を修正
- テストファイル: 型アノテーションと type: ignore を追加

+++ src/my_lib/pil_util.py
@@ -66,7 +66,7 @@ def draw_text(  # noqa: PLR0913
     color: str = "#000",
     stroke_width: int = 0,
     stroke_fill: str | tuple[int, int, int, int] | None = None,
-) -> tuple[int, int]:
+) -> tuple[float, float]:
     text_line_list = text.split("\n")
 
     pos_x, next_pos_y = pos
